% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_examples_in_rd.R
\name{has_examples_in_rd}
\alias{has_examples_in_rd}
\title{Checks if an R documentation file supplies one or more examples}
\usage{
has_examples_in_rd(rd_filename)
}
\arguments{
\item{rd_filename}{name of an R documentation (.Rd) file. These are located in a package its 'man' folder}
}
\value{
TRUE or FALSE
}
\description{
Checks if an R documentation file supplies one or more examples
}
\examples{
  filename <- paste0(getwd(), "/man/has_examples_in_rd.Rd")
  if (file.exists(filename)) {
    testit::assert(has_examples_in_rd(filename))
  }
}
\author{
Richel Bilderbeek
}

