% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_tags.R
\name{rd_tags}
\alias{rd_tags}
\title{Copy of tools::RdTags. Because tools::RdTags
is not an exported function, I duplicated its code.}
\usage{
rd_tags(rd)
}
\arguments{
\item{rd}{a parsed .Rd file, as produced by tools::parse_Rd}
}
\value{
the tags in the parsed .Rd file
}
\description{
Copy of tools::RdTags. Because tools::RdTags
is not an exported function, I duplicated its code.
}
\examples{
  filename <- find_path("f_with_examples.Rd")
  doc <- tools::parse_Rd(rd_filename)
  testit::assert(class(doc) == "Rd")
  tags <- rd_tags(doc)
  testit::assert("\\\\examples" \%in\% tags)
}
\author{
Richel Bilderbeek
}

